changed end points and refactor
